task: #211 Combine Simulation and SimulationAlt
